l10n_et_hr: Improve display name in both HR and public views
