+++ l10n_et_hr/__manifest__.py
@@ -16,6 +16,7 @@
         "ethiopic_calendar",
     ],
     "data": [
+        "views/hr_employee_public_views.xml",
         "views/hr_employee_views.xml",
         "views/res_config_settings_views.xml",
     ],
