+++ l10n_et_hr/models/__init__.py
@@ -4,5 +4,7 @@
 
 from . import hr_department
 from . import hr_employee
+from . import hr_employee_base
+from . import hr_employee_public
 from . import hr_job
 from . import res_config_settings
