+++ l10n_et_hr/views/res_config_settings_views.xml
@@ -7,37 +7,49 @@
         <field name="priority" eval="80" />
         <field name="inherit_id" ref="hr.res_config_settings_view_form" />
         <field name="arch" type="xml">
-            <xpath expr="//div[hasclass('app_settings_block')]" position="inside">
-                <div class="col-12 col-lg-6 o_setting_box" id="l10n_hr_employee_name">
-                    <div class="o_setting_left_pane">
-                        <field name="use_ethiopic_name" />
-                    </div>
-                    <div class="o_setting_right_pane">
-                        <label for="use_ethiopic_name" />
-                        <div class="text-muted">
-                                By default display employee names in the Amharic script
+            <xpath expr="//div[@data-key='hr']" position="inside">
+                <h2>Ethiopian Localization</h2>
+                <div
+                    class="row mt16 o_settings_container"
+                    name="employee_rights_setting_container"
+                >
+                    <div
+                        class="col-12 col-lg-6 o_setting_box"
+                        id="l10n_hr_employee_name"
+                    >
+                        <div class="o_setting_left_pane">
+                            <field name="use_ethiopic_name" />
+                        </div>
+                        <div class="o_setting_right_pane">
+                            <label for="use_ethiopic_name" />
+                            <div class="text-muted">
+                                    By default display employee names in the Amharic script
+                            </div>
                         </div>
                     </div>
-                </div>
-                <div class="col-12 col-lg-6 o_setting_box" id="l10n_hr_department_name">
-                    <div class="o_setting_left_pane">
-                        <field name="use_ethiopic_department" />
-                    </div>
-                    <div class="o_setting_right_pane">
-                        <label for="use_ethiopic_department" />
-                        <div class="text-muted">
-                                By default display department names in the Amharic script
+                    <div
+                        class="col-12 col-lg-6 o_setting_box"
+                        id="l10n_hr_department_name"
+                    >
+                        <div class="o_setting_left_pane">
+                            <field name="use_ethiopic_department" />
+                        </div>
+                        <div class="o_setting_right_pane">
+                            <label for="use_ethiopic_department" />
+                            <div class="text-muted">
+                                    By default display department names in the Amharic script
+                            </div>
                         </div>
                     </div>
-                </div>
-                <div class="col-12 col-lg-6 o_setting_box" id="l10n_hr_job_name">
-                    <div class="o_setting_left_pane">
-                        <field name="use_ethiopic_job" />
-                    </div>
-                    <div class="o_setting_right_pane">
-                        <label for="use_ethiopic_job" />
-                        <div class="text-muted">
-                                By default display job titles in the Amharic script
+                    <div class="col-12 col-lg-6 o_setting_box" id="l10n_hr_job_name">
+                        <div class="o_setting_left_pane">
+                            <field name="use_ethiopic_job" />
+                        </div>
+                        <div class="o_setting_right_pane">
+                            <label for="use_ethiopic_job" />
+                            <div class="text-muted">
+                                    By default display job titles in the Amharic script
+                            </div>
                         </div>
                     </div>
                 </div>
